+++ app/screen/OnboardingScreen.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useCallback } from 'react'
 import {View,Text,StyleSheet,useWindowDimensions} from 'react-native'
 import Onboarding from 'react-native-onboarding-swiper'
 import LottieView from 'lottie-react-native';
@@ -22,10 +22,9 @@ const OnboardingScreen = () => {
     const theme = useTheme()
     const {width} = useWindowDimensions()
 
-    const handleDone = async () => {
+    const handleDone = useCallback(async () => {
         await AsyncStorage.setItem('onboarded','1')
-        navigation.navigate('Drawer',{screen:'Home'})
-    }
+    },[])
   return (
     <View style={styles.screen}>
     <Onboarding
@@ -34,7 +33,7 @@ const OnboardingScreen = () => {
         containerStyles={{paddingHorizontal:15}}
         pages={[
             {
-            backgroundColor: '#00bf63',
+            backgroundColor:theme.colors.background,
             titleStyles:{
                 fontFamily:color.textFont
             },
@@ -50,7 +49,7 @@ const OnboardingScreen = () => {
             subtitle: 'Your journey to becoming a better developer starts here!'
             },
             {
-                backgroundColor: '#fff',
+                backgroundColor: theme.colors.background,
                 titleStyles:{
                     fontFamily:color.textFont
                 },
@@ -66,7 +65,7 @@ const OnboardingScreen = () => {
                 subtitle: 'Turn your side projects into real-world applications with Foundry'
                 },
             {
-            backgroundColor: '#0097b2',
+            backgroundColor:theme.colors.background,
             titleStyles:{
                 fontFamily:color.textFont
             },
@@ -82,7 +81,7 @@ const OnboardingScreen = () => {
             subtitle: 'From concepts to code, Foundry is here to help you make it happen'
             },
             {
-                backgroundColor: color.grey,
+                backgroundColor: theme.colors.background,
                 titleStyles:{
                     fontFamily:color.textFont
                 },
@@ -98,7 +97,7 @@ const OnboardingScreen = () => {
                 subtitle: 'Collaborate with developers from all skill levels and backgrounds'
                 },
             {
-            backgroundColor:"#fff",
+            backgroundColor:theme.colors.background,
             titleStyles:{
                 fontFamily:color.textFont
             },
