+++ config/color.ts
@@ -1,3 +1,104 @@
+import {  MD3LightTheme as DefaultTheme,MD3DarkTheme as DarkTheme, } from 'react-native-paper';
+export const lightTheme = {
+    ...DefaultTheme,
+    colors: {
+      ...DefaultTheme.colors,
+      primary: 'rgb(0, 130, 242)',
+      onPrimary: 'rgb(5, 126, 230)',
+      primaryContainer: 'rgb(220, 225, 255)',
+      onPrimaryContainer: 'rgb(0, 22, 78)',
+      secondary: 'rgb(141, 65, 137)',
+      onSecondary: 'rgb(255, 255, 255)',
+      secondaryContainer: 'rgb(255, 215, 246)',
+      onSecondaryContainer: 'rgb(56, 0, 57)',
+      tertiary: 'rgb(0, 0, 0)',
+      onTertiary: 'rgb(128, 128, 128)',
+      tertiaryContainer: 'rgb(0, 0, 0)',
+      onTertiaryContainer: 'rgb(44, 18, 42)',
+      error: 'rgb(186, 26, 26)',
+      onError: 'rgb(255, 255, 255)',
+      errorContainer: 'rgb(255, 218, 214)',
+      onErrorContainer: 'rgb(65, 0, 2)',
+      background: 'rgb(255, 255, 255)',
+      onBackground: 'rgb(255, 255, 255)',
+      surface: 'rgb(254, 251, 255)',
+      onSurface: 'rgb(27, 27, 31)',
+      surfaceVariant: 'rgb(226, 225, 236)',
+      onSurfaceVariant: 'rgb(69, 70, 79)',
+      outline: 'rgb(118, 118, 128)',
+      outlineVariant: 'rgb(198, 198, 208)',
+      shadow: 'rgb(0, 0, 0)',
+      scrim: 'rgb(0, 0, 0)',
+      inverseSurface: 'rgb(48, 48, 52)',
+      inverseOnSurface: 'rgb(242, 240, 244)',
+      inversePrimary: 'rgb(182, 196, 255)',
+      elevation: {
+        level0: 'transparent',
+        level1: 'rgb(245, 243, 251)',
+        level2: 'rgb(239, 238, 248)',
+        level3: 'rgb(233, 233, 246)',
+        level4: 'rgb(231, 232, 245)',
+        level5: 'rgb(227, 228, 243)'
+      },
+      surfaceDisabled: 'rgba(27, 27, 31, 0.12)',
+      onSurfaceDisabled: 'rgba(27, 27, 31, 0.38)',
+      backdrop: 'rgb(47, 48, 56, 0.4)',
+      backgroundContainer: 'rgb(190, 233, 255)',
+      onBackgroundContaine: 'rgb(0, 31, 42)',
+        },
+  }
+export const darkTheme = {
+    ...DarkTheme,
+    colors: {
+      ...DarkTheme.colors,
+      text:'#fff',
+      primary: 'rgb(0, 130, 242)',
+      onPrimary: 'rgb(247, 247, 247)',
+      primaryContainer: 'rgb(38, 65, 145)',
+      onPrimaryContainer: 'rgb(220, 225, 255)',
+      secondary: 'rgb(84, 95, 113)',
+      onSecondary: 'rgb(69, 90, 100)',
+      secondaryContainer: 'rgb(114, 40, 111)',
+      onSecondaryContainer: 'rgb(255, 215, 246)',
+      tertiary: 'rgb(250, 250, 250)',
+      onTertiary: 'rgb(128, 128, 128)',
+      tertiaryContainer: 'rgb(0, 0, 0)',
+      onTertiaryContainer: 'rgb(255, 215, 246)',
+      error: 'rgb(255, 180, 171)',
+      onError: 'rgb(105, 0, 5)',
+      errorContainer: 'rgb(147, 0, 10)',
+      onErrorContainer: 'rgb(255, 180, 171)',
+      background: 'rgb(0, 0, 0)',
+      onBackground: 'rgb(51, 51, 51)',
+      surface: 'rgb(27, 27, 31)',
+      onSurface: 'rgb(228, 225, 230)',
+      surfaceVariant: 'rgb(69, 70, 79)',
+      onSurfaceVariant: 'rgb(198, 198, 208)',
+      outline: 'rgb(143, 144, 154)',
+      outlineVariant: 'rgb(69, 70, 79)',
+      shadow: 'rgb(0, 0, 0)',
+      scrim: 'rgb(0, 0, 0)',
+      inverseSurface: 'rgb(228, 225, 230)',
+      inverseOnSurface: 'rgb(48, 48, 52)',
+      inversePrimary: 'rgb(64, 89, 170)',
+      elevation: {
+        level0: 'transparent',
+        level1: 'rgb(35, 35, 42)',
+        level2: 'rgb(39, 41, 49)',
+        level3: 'rgb(44, 46, 56)',
+        level4: 'rgb(46, 47, 58)',
+        level5: 'rgb(49, 51, 62)'
+      },
+      surfaceDisabled: 'rgba(228, 225, 230, 0.12)',
+      onSurfaceDisabled: 'rgba(228, 225, 230, 0.38)',
+      backdrop: 'rgb(128, 128, 128)',
+      backgroundContainer: 'rgb(0, 77, 100)',
+      onBackgroundContainer: 'rgb(190, 233, 255)',
+      onText: 'rgb(0, 52, 79)',
+      textContainer: 'rgb(0, 75, 113)',
+      onTextContainer: 'rgb(203, 230, 255)'
+        }
+      }
  const color = {
     light:"#f8f4f4",
     primary: '#E0FBFC',
